Backend/requirements.txt
flask
python-dotenv
openai
opencv-contrib-python
ultralytics
numpy
requests